Rename CamOpsRatingV2 package to CamOpsReputation
Renamed the CamOpsRatingV2 package and its classes to CamOpsReputation for better clarity and alignment with the new naming conventions. Updated all relevant import statements to reflect these changes.
